extra_css_urls and extra_js_urls for render_template()
